+++ packages/spotlight/src/spotlight.js
@@ -231,7 +231,7 @@ const Spotlight = (function () {
 			const nextContainerIds = getContainersForNode(next);
 
 			// prevent focus if 5-way is being held and the next element would change containers
-			if (_5WayKeyHold && last(nextContainerIds) !== currentContainerId) {
+			if (_5WayKeyHold && nextContainerIds.indexOf(currentContainerId) < 0) {
 				return false;
 			}